Set game as DPI-aware on Windows
This prevents Windows from performing bilinear scaling in addition to
whatever the game is doing. SDL_WINDOW_ALLOW_HIGHDPI should do this on
all platforms, but it's only implemented for macOS, Wayland (X11 doesn't
have a concept of DPI-awareness), and a few more esoteric ones. Windows
has a few ways of doing it, but SetProcessDPIAware is the easiest. The
docs say that this can cause unexpected behavior, but my understanding
is that this relates to multithreading, and in this case nothing should
be able to go wrong.